Working on Initial Scene Transitions
Created a main menu scene, and am trying to transition from that to gameplay. I have separated the gameplay scenes into the backgrounds, and a persistentgameplay scene. The transition from main menu to the game is buggy, partly because of coroutines, and maybe even because multiple scenes means multiple cameras.